requirements.txt
customtkinter
Pillow
faker
CTkSpinbox
CTkMessagebox